requirements.txt
Django==2.1
djangorestframework==3.8.2
coreapi==2.3.3
django-heroku==0.3.1 
gunicorn==19.9.0
whitenoise==3.3.1
short-url==1.2.2
